refactor: split credential key routes
